Factor out caching in a separate module
